+++ src-tauri/src/lib.rs
@@ -20,6 +20,7 @@ pub fn run() {
         .plugin(tauri_plugin_clipboard_manager::init())
         .plugin(tauri_plugin_process::init())
         .plugin(tauri_plugin_os::init())
+        .plugin(tauri_plugin_updater::Builder::new().build())
         .register_uri_scheme_protocol("local-file", |_ctx, req| {
             storage::LocalFileProtocol::handle_request(req)
         })